+++ app/src/main/java/com/dimas/goodnews/presentation/fragments/StartFragment.kt
@@ -11,10 +11,10 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.dimas.goodnews.R
 import com.dimas.goodnews.data.network.models.Article
 import com.dimas.goodnews.databinding.FragmentStartBinding
-import com.dimas.goodnews.domain.utils.Resource
+import com.dimas.goodnews.utils.Resource
 import com.dimas.goodnews.presentation.MainActivity
 import com.dimas.goodnews.presentation.adapters.ArticleAdapter
-import com.dimas.goodnews.presentation.viewmodels.StartViewModel
+import com.dimas.goodnews.presentation.viewmodels.NewsViewModel
 import kotlinx.android.synthetic.main.fragment_start.*
 import kotlin.random.Random
 
@@ -23,7 +23,7 @@ class StartFragment : Fragment() {
     private val binding get() = _binding ?: throw RuntimeException("FragmentStartBinding == null")
 
     lateinit var newsAdapter: ArticleAdapter
-    lateinit var viewModel: StartViewModel
+    lateinit var viewModel: NewsViewModel
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
